completionPage.py -> LlamaPage.py, remove legacy function
