+++ .gitignore
@@ -130,8 +130,5 @@ dmypy.json
 
 # pyqt_openai
 pyqt_openai/pyqt_openai.ini
-pyqt_openai/conv.json
-pyqt_openai/conv_history.json
 pyqt_openai/conv.db
-pyqt_openai/a.py
 pyqt_openai/test/
